Moved the free port determination methods to a separate file. Added tests.
